+++ src/Google/Api/Ads/AdWords/Util/XmlDeserializer.php
@@ -88,14 +88,20 @@ class XmlDeserializer {
           $typeHint = ($elementObject !== null)
               ? self::GetTypeHint(get_class($elementObject), $childName)
               : null;
+
+          // Type hint is sometimes suffixed by [] to indicate that the member
+          // variable is supposed to be an array.
+          if (substr($typeHint, -strlen('[]')) === '[]') {
+            if (array_key_exists($childName, $result) === false) {
+              $result[$childName] = array();
+            }
+            $typeHint = substr($typeHint, 0, -strlen('[]'));
+          }
           // Call this function recursively on the child node.
           $value = self::ConvertElementToObject($childNode, $typeHint);
           if (!isset($result[$childName])) {
             $result[$childName] = $value;
           } else {
-            if (!is_array($result[$childName])) {
-              $result[$childName] = array($result[$childName]);
-            }
             $result[$childName][] = $value;
           }
         }
@@ -226,11 +232,9 @@ class XmlDeserializer {
       $prop = $clazz->getProperty($propertyName);
       if (preg_match('/@var\s*(.+)\s/', $prop->getDocComment(),
           $annotations)) {
-        // The type hints are sometimes prefixed with "tns" or suffixed
-        // by [] to indicate that the variable stores array.
-        // They are stripped here to extract only the type names.
+        // The type hints are sometimes prefixed with "tns", which is stripped
+        // here to extract only the type names.
         $typeHint = preg_replace('/^tns/', '', $annotations[1]);
-        $typeHint = preg_replace('/\[\]$/', '', $typeHint);
       }
     }
     return $typeHint;